change math hit score to float type.

+++ search/math-expr-sim.h
@@ -12,7 +12,7 @@ struct math_extra_score_arg;
 struct math_expr_score_res {
 	doc_id_t  doc_id;
 	exp_id_t  exp_id;
-	uint32_t  score;
+	float     score;
 #ifdef MATH_SLOW_SEARCH
 	/* postiional info */
 	uint64_t  qmask[MAX_MTREE];
